pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brix"
version = "7.41.4"
description = "Generic process orchestrator for Claude Code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.20",
    "httpx>=0.28",
    "pydantic>=2.12",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "click>=8.3",
    "starlette>=0.40",
    "uvicorn>=0.30",
    "fints>=3.0",
    "mt-940>=4.0",
    "mistralai>=1.0,<2.0",
    "psycopg2-binary>=2.9",
    "cryptography>=42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
brix = "brix.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/brix"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src", "tests"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.mypy]
python_version = "3.12"
strict = true
